requirements.txt
Flask==3.1.0
tensorflow-cpu==2.18.0
numpy==2.0.2
gunicorn==20.1.0